refactor: 예약 가능 시간 조회 로직 ThemeService로 이동
