+++ src/main/java/roomescape/controller/ThemeController.java
@@ -6,7 +6,6 @@ import java.time.LocalDate;
 import java.util.List;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -17,7 +16,6 @@ import org.springframework.web.bind.annotation.RestController;
 import roomescape.controller.dto.ThemeRequest;
 import roomescape.controller.dto.ThemeResponse;
 import roomescape.domain.Theme;
-import roomescape.service.ReservationService;
 import roomescape.service.ThemeService;
 import roomescape.service.dto.TimeAvailabilityDto;
 
@@ -26,11 +24,9 @@ import roomescape.service.dto.TimeAvailabilityDto;
 public class ThemeController {
 
     private final ThemeService themeService;
-    private final ReservationService reservationService;
 
-    public ThemeController(ThemeService themeService, ReservationService reservationService) {
+    public ThemeController(ThemeService themeService) {
         this.themeService = themeService;
-        this.reservationService = reservationService;
     }
 
     @GetMapping
@@ -58,7 +54,7 @@ public class ThemeController {
 
     @GetMapping("/{id}/times")
     public List<TimeAvailabilityDto> getAvailableTimes(@PathVariable Long id, @RequestParam("date") LocalDate date) {
-        return reservationService.findAvailableTime(id, date);
+        return themeService.findAvailableTime(id, date);
     }
 
     @GetMapping("/popular")
