+++ src/main/java/roomescape/service/ThemeService.java
@@ -5,17 +5,27 @@ import java.util.List;
 import java.util.NoSuchElementException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import roomescape.domain.Reservation;
+import roomescape.domain.ReservationTime;
 import roomescape.domain.Theme;
+import roomescape.repository.ReservationRepository;
+import roomescape.repository.ReservationTimeRepository;
 import roomescape.repository.ThemeRepository;
+import roomescape.service.dto.TimeAvailabilityDto;
 
 @Service
 @Transactional(readOnly = true)
 public class ThemeService {
 
     private final ThemeRepository themeRepository;
+    private final ReservationRepository reservationRepository;
+    private final ReservationTimeRepository reservationTimeRepository;
 
-    public ThemeService(ThemeRepository themeRepository) {
+    public ThemeService(ThemeRepository themeRepository, ReservationRepository reservationRepository,
+                        ReservationTimeRepository reservationTimeRepository) {
         this.themeRepository = themeRepository;
+        this.reservationRepository = reservationRepository;
+        this.reservationTimeRepository = reservationTimeRepository;
     }
 
     public List<Theme> findAll() {
@@ -42,5 +52,22 @@ public class ThemeService {
         LocalDate startDate = LocalDate.now().minusWeeks(1);
         LocalDate endDate = startDate.plusDays(6);
         return themeRepository.findPopular(startDate, endDate, 10);
+    }
+
+    public List<TimeAvailabilityDto> findAvailableTime(Long themeId, LocalDate date) {
+        List<ReservationTime> times = reservationTimeRepository.findAll();
+        List<Reservation> reservations = reservationRepository.findReservationsByThemeAndDate(themeId, date);
+
+        return times.stream()
+                .map(time -> new TimeAvailabilityDto(
+                        time,
+                        isAvailable(time, reservations)
+                ))
+                .toList();
+    }
+
+    private boolean isAvailable(ReservationTime time, List<Reservation> reservations) {
+        return reservations.stream()
+                .noneMatch(reservation -> time.equals(reservation.getTime()));
     }
 }
